refactor: コードレビュー指摘修正（MR-1, MR-2, RC-1, RC-4, SG-1）
- MR-1: applyDeflectionBias を ai.ts → physics.ts に移動（entities→ai の依存逆転を解消）
- MR-2: テストの require() 動的インポートを静的 import に変更（型安全性向上）
- RC-1: applyDefenseStyle にランダムオフセット（±20px）を追加（機械的な固定ポジションを回避）
- RC-4: deflectionBias テストの assertion を角度比較で具体化
- SG-1: TEAM_ROLE_AGGRESSIVENESS の型を Record<string, number> → Record<TeamRole, number> に限定
