+++ src/features/air-hockey/core/ai.test.ts
@@ -1,4 +1,5 @@
-import { CpuAI, applyAdaptability } from './ai';
+import { CpuAI, applyAdaptability, shouldRecalculateTarget } from './ai';
+import { applyDeflectionBias } from './physics';
 import { EntityFactory } from './entities';
 import { CONSTANTS } from './constants';
 import type { AiBehaviorConfig } from './story-balance';
@@ -414,25 +415,24 @@ describe('sidePreference', () => {
 
 describe('deflectionBias', () => {
   it('deflectionBias > 0 でパックの反射方向が壁方向に偏る', () => {
-    // deflectionBias の効果は resolveMalletPuckOverlap 内で適用される
-    // ここでは applyDeflectionBias 関数を直接テスト
-    const { applyDeflectionBias } = require('./ai');
-    // 真上方向の法線 (0, -1)
+    // 真上方向の法線 (0, -1) に正のバイアス → 壁方向（水平）に傾く
     const result = applyDeflectionBias(0, -1, 0.5);
-    // bias > 0: 壁方向（水平）に傾く → nx の絶対値が増加
-    expect(Math.abs(result.nx)).toBeGreaterThan(0);
+    // 元の角度は -π/2（真上）、バイアスで水平方向に傾くので nx の絶対値が増加
+    const originalAngle = Math.atan2(-1, 0);
+    const biasedAngle = Math.atan2(result.ny, result.nx);
+    expect(biasedAngle).toBeGreaterThan(originalAngle);
   });
 
   it('deflectionBias < 0 でパックの反射方向がゴール方向に偏る', () => {
-    const { applyDeflectionBias } = require('./ai');
-    // 斜め方向の法線
+    // 斜め方向の法線 (0.7, -0.7) に負のバイアス → ゴール方向（垂直）に傾く
+    const originalAngle = Math.atan2(-0.7, 0.7);
     const result = applyDeflectionBias(0.7, -0.7, -0.5);
-    // bias < 0: ゴール方向（垂直）に傾く
-    expect(result.ny).toBeDefined();
+    const biasedAngle = Math.atan2(result.ny, result.nx);
+    // 負のバイアスで角度がより垂直方向（小さい値）に傾く
+    expect(biasedAngle).toBeLessThan(originalAngle);
   });
 
   it('deflectionBias = 0 では法線が変化しない', () => {
-    const { applyDeflectionBias } = require('./ai');
     const result = applyDeflectionBias(0, -1, 0);
     expect(result.nx).toBeCloseTo(0, 5);
     expect(result.ny).toBeCloseTo(-1, 5);
@@ -443,22 +443,18 @@ describe('deflectionBias', () => {
 
 describe('shouldRecalculateTarget', () => {
   it('パック方向転換時に reactionDelay 経過後は再計算を許可する', () => {
-    const { shouldRecalculateTarget } = require('./ai');
     expect(shouldRecalculateTarget(0, 100, 50, true)).toBe(true);
   });
 
   it('パック方向転換時に reactionDelay 未経過なら再計算しない', () => {
-    const { shouldRecalculateTarget } = require('./ai');
     expect(shouldRecalculateTarget(0, 30, 50, true)).toBe(false);
   });
 
   it('パック方向転換がなければ再計算しない', () => {
-    const { shouldRecalculateTarget } = require('./ai');
     expect(shouldRecalculateTarget(0, 1000, 50, false)).toBe(false);
   });
 
   it('reactionDelay=0 ではパック方向転換時に即座に再計算', () => {
-    const { shouldRecalculateTarget } = require('./ai');
     expect(shouldRecalculateTarget(0, 0, 0, true)).toBe(true);
   });
 });
